Move fetch to utils
