+++ templates/vue/config/_entrypoint.js
